2023/ResQRs/QR-webapp/requirements.txt
Flask==1.1.2
Jinja2==2.11.3
jinja2-time==0.2.0
python-dateutil==2.7.3
requests==2.22.0
Werkzeug==1.0.1
markupsafe==2.0.1
itsdangerous==2.0.1
pandas==1.3.0
segno==1.5.2
Pillow